refactor:  profile 모달, group home 페이지 리팩토링
