+++ src/components/modal/profile-modal/profile-modal.tsx
@@ -3,9 +3,10 @@ import Modal from '@components/modal/modal';
 import ModalButton from '@components/modal/modal-button';
 import ModalHeader from '@components/modal/modal-header';
 import InvitedList from '@components/modal/profile-modal/invited-list';
-import UserInfo from '@components/modal/profile-modal/user-info';
+import NicknameInput from '@components/modal/profile-modal/nickname-input';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
+import ProfileImageInput from './profile-image-input';
 
 interface GroupModalProps {
   children: React.ReactNode;
@@ -33,10 +34,11 @@ const ProfileModal = ({ children }: GroupModalProps) => {
       <Modal isOpen={isOpen} onClose={handleModalClose}>
         <ModalHeader headerTitle="내 정보" onClose={handleCloseButtonClick} />
 
-        <S.MyInfoContent>
-          <UserInfo />
+        <div>
+          <ProfileImageInput />
+          <NicknameInput />
           <InvitedList />
-        </S.MyInfoContent>
+        </div>
 
         <S.ModalFooter>
           <S.WithdrawButton>탈퇴하기</S.WithdrawButton>
@@ -52,10 +54,6 @@ const ProfileModal = ({ children }: GroupModalProps) => {
 export default ProfileModal;
 
 const S = {
-  MyInfoContent: styled.div`
-    margin-top: 0px;
-  `,
-
   ModalFooter: styled.div`
     margin-top: 15px;
     width: 100%;
